A bit of clean-up and docs.
