+++ data/.gitignore
@@ -1,1 +1,1 @@
-*-results
+results-*
